Better iteration

+++ app/main.py
@@ -184,22 +184,25 @@ def delete_images_from_meural_if_needed(meural_token, icloud_album_id, album_che
     logger.info(f"Preparing to delete {len(since_been_deleted_checksums)} items from Meural")
     for checksum in since_been_deleted_checksums:
         if checksum in Metadata.db[icloud_album_id]:
-            for playlist_name in Metadata.db[icloud_album_id][checksum]:
-                playlist_data = Metadata.db[icloud_album_id][checksum][playlist_name]
-                image_filename = playlist_data['filename']
-                try:
-                    meural.delete_image(meural_token, playlist_data['meural_id'])
-                    Metadata.delete_item(icloud_album_id, checksum, playlist_name)
+            # Playlists may be deleted at this point, so grab what they were and check if they still exist
+            playlist_names = list(Metadata.db[icloud_album_id][checksum].keys())
+            for playlist_name in playlist_names:
+                if playlist_name in Metadata.db[icloud_album_id][checksum]:
+                    playlist_data = Metadata.db[icloud_album_id][checksum][playlist_name]
+                    image_filename = playlist_data['filename']
+                    try:
+                        meural.delete_image(meural_token, playlist_data['meural_id'])
+                        Metadata.delete_item(icloud_album_id, checksum, playlist_name)
 
-                    potential_image_location = f"{IMAGE_DIR}/not_uploaded/{image_filename}"
-                    for potential_location in (potential_image_location, potential_image_location.replace('/not_uploaded/', '/uploaded/')):
-                        if os.path.isfile(potential_location):
-                            os.remove(potential_location)
-                            logger.info(f"[✓] Successfully deleted {image_filename} from local storage")
-                            break
-                    logger.info(f"[✓] Successfully deleted {image_filename} from Meural")
-                except Exception as e:
-                    logger.error(f"[X] Failed to delete {image_filename} from Meural: {e}")
+                        potential_image_location = f"{IMAGE_DIR}/not_uploaded/{image_filename}"
+                        for potential_location in (potential_image_location, potential_image_location.replace('/not_uploaded/', '/uploaded/')):
+                            if os.path.isfile(potential_location):
+                                os.remove(potential_location)
+                                logger.info(f"[✓] Successfully deleted {image_filename} from local storage")
+                                break
+                        logger.info(f"[✓] Successfully deleted {image_filename} from Meural")
+                    except Exception as e:
+                        logger.error(f"[X] Failed to delete {image_filename} from Meural: {e}")
         else:
             logger.warning(f"Checksum {checksum} not found in metadata for album id {icloud_album_id}")
     Metadata.clean_db(icloud_album_id)
